Minor changes as refactoring consequences
