+++ client/src/components/CreateGame/Lecturer/Lecturer.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react"
 import "./Lecturer.css"
 import { useNavigate } from "react-router-dom"
 import socket from "../../../socket"
-import useWindowDimensions from "../../RaceThemes/TrainTheme/Tracks/WindowDimensions"
-import StationDisplay from "../../RaceThemes/TrainTheme/StationDisplay/StationDisplay"
+import useWindowDimensions from "../../RaceThemes/Tracks/WindowDimensions"
+import StationDisplay from "../../RaceThemes/StationDisplay/StationDisplay"
 import CheckPoint from "./LeaderBoard/CheckPoint"
 import LeaderBoard from "./LeaderBoard/LeaderBoard"
 import QuestionStatistics from "./QuestionStatistics/QuestionStatistics"
