+++ client/src/components/RaceThemes/Ghosts/Ghosts.css
@@ -2,26 +2,4 @@
     opacity: 0.7;
 }
 
-.ghost > .name,
-.best-ghost > .name {
-    position: absolute;
-    text-align: right;
-    font-size: 16px;
-    font-weight: 500;
 
-    padding-left: 5px;
-    margin-left: 56px;
-    margin-top: 5px;
-    border-bottom-width: 3px;
-    border-bottom-style: solid;
-    background-color: white;
-}
-
-.highest-score {
-    position: absolute;
-    top: 60px;
-    font-size: 18px;
-    color: #ffd700;
-    font-weight: 600;
-    opacity: 1;
-}
